requirements.txt
openai>=1.0.0
pandas
openpyxl
xlsxwriter
python-dotenv
streamlit